+++ pciw/units.py
@@ -1,55 +1,14 @@
-import os
-import glob
-import platform
-
-# ! Функции получения тега системы
-def systag() -> str:
-    return f"{platform.system()}-{platform.machine()}"
-
-# ! Для функций
-NONE_TYPE_EXCEPTIONS = [
-    "tobefilled", "default", "n/a",
-    "[notsupported]", "[n/a]",
-    "tobefilledbyo.e.m."
+NONE_VALUES = [
+    "tobefilledbyo.e.m.",
+    "tobefilled",
+    "default",
+    "n/a",
+    "[notsupported]",
+    "[n/a]",
 ]
 
-# ! Поиск nvidia-smi (WINDOWS ONLY)
-WINDOWS_NVIDIA_SMI_PATH = "nvidia-smi"
-WINDOWS_NVIDIA_SMI_POSSIBLE_PATHS = [
-    "{sd}\\Windows\\System32\\DriverStore\\FileRepository\\*\\nvidia-smi.exe",
-    "{sd}\\Program Files\\NVIDIA Corporation\\NVSMI\\*\\nvidia-smi.exe"
-]
-
-if platform.system() == "Windows":
-    system_drive = os.getenv("SystemDrive")
-    for possible_path in WINDOWS_NVIDIA_SMI_POSSIBLE_PATHS:
-        gwnsmi = glob.glob(possible_path.format(sd=system_drive), recursive=True)
-        if len(gwnsmi) != 0:
-            WINDOWS_NVIDIA_SMI_PATH = gwnsmi[0]
-            break
-    del system_drive
-
-# ! Определение путей
-NVIDIA_SMI_PATH_SUPPORTED = {
-    "Windows-AMD64": WINDOWS_NVIDIA_SMI_PATH,
-    "Windows-x86_64": WINDOWS_NVIDIA_SMI_PATH,
-    "Linux-x86_64": "nvidia-smi"
-}
-T_CPU_PATH_SUPPORTED = {
-    "Windows-AMD64": os.path.join(os.path.dirname(__file__), "data\\t_cpu\\Windows-x86_64\\parser.py"),
-    "Windows-x86_64": os.path.join(os.path.dirname(__file__), "data\\t_cpu\\Windows-x86_64\\parser.py")
-}
-T_CPU_PATH_ADMIN_SUPPORTED = {
-    "Windows-AMD64": os.path.join(os.path.dirname(__file__), "data\\t_cpu\\Windows-x86_64\\t_cpu.exe"),
-    "Windows-x86_64": os.path.join(os.path.dirname(__file__), "data\\t_cpu\\Windows-x86_64\\t_cpu.exe")
-}
-
-NVIDIA_SMI_PATH: str = NVIDIA_SMI_PATH_SUPPORTED.get(systag(), None)
-T_CPU_PATH: str = T_CPU_PATH_SUPPORTED.get(systag(), None)
-T_CPU_PATH_ADMIN: str = T_CPU_PATH_ADMIN_SUPPORTED.get(systag(), None)
-
-# ! Для определения значений
-class NT:
+# ! NT Types
+class WindowsTypes:
     MEMORY_TYPE = {
         0: None,
         1: '<UNKNOWN>',